refactor and add some integration tests for postprocessors
